Cambiando Longs por Integers

+++ src/main/java/bo/edu/uto/encuestadea/controladores/ComisionController.java
@@ -36,7 +36,7 @@ public class ComisionController {
 
 		HashMap modelo = new HashMap();
 
-		Integer id_usuario = (Long) session.getAttribute("__id_usuario");
+		Integer id_usuario = (Integer) session.getAttribute("__id_usuario");
 		String apodo = (String) session.getAttribute("__apodo");
 		modelo.put("logout", id_usuario == null);
 
@@ -51,9 +51,9 @@ public class ComisionController {
 	@ResponseBody
 	public Map guardar(HttpSession hs, IntegranteComision integrante) throws IllegalAccessException, IllegalArgumentException, InvocationTargetException, NoSuchMethodException {
 		Map modelo = new HashMap();
-		Integer id_usuario = (Long) hs.getAttribute("__id_usuario");
+		Integer id_usuario = (Integer) hs.getAttribute("__id_usuario");
 		integrante.setId_usuario(id_usuario);
-		integrante.setId_comision(0L);
+		integrante.setId_comision(0);
 		integranteComisionMapa.insertarIntegrante(integrante);
 		modelo.put("resultado", true);
 		return modelo;
@@ -83,11 +83,11 @@ public class ComisionController {
 	public boolean modificar(IntegranteComision integranteComision, HttpSession hs) {
 		Map modelo = new HashMap();
 		// Verificando si el Usuario sigue autentificado.
-		Integer id_usuario = (Long) hs.getAttribute("__id_usuario");
+		Integer id_usuario = (Integer) hs.getAttribute("__id_usuario");
 		modelo.put("logout", id_usuario == null);
 
 		integranteComision.setId_usuario(id_usuario);
-		integranteComision.setId_comision(0L);
+		integranteComision.setId_comision(0);
 		int cantidadModificados = integranteComisionMapa.modificarInegranteComision(integranteComision);
 		if (cantidadModificados > 0) {
 			return true;
